refactor(stage): split watchdog into ndi_watchdog.rs, extract WS read loop — size caps
ndi_video.rs hit the 1000-line hard cap (1058) after the frame-based
watchdog rework; the watchdog + codec-mode state + beacons move to
ndi_watchdog.rs (no behavior change). spawn_stage_ws exceeded the 120-line
cap (138); the read loop and text-frame dispatch are now free fns.
